+++ sistema/src/main/java/com/perdi2enlauni/sistema/model/Usuario.java
@@ -1,0 +1,80 @@
+package com.perdi2enlauni.sistema.model;
+
+import jakarta.persistence.*;
+
+@MappedSuperclass
+public class Usuario {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    private String nombre;
+    private String correo;
+    private String dni;
+    private String contrasenia;
+
+    @Column(name = "rol")
+    private String rol; // Role attribute
+
+    public Usuario() {
+    }
+
+    public Usuario(int id, String nombre, String correo, String dni, String contrasenia, String rol) {
+        this.id = id;
+        this.nombre = nombre;
+        this.correo = correo;
+        this.dni = dni;
+        this.contrasenia = contrasenia;
+        this.rol = rol;
+    }
+
+    // Getters and Setters
+    public int getId() {
+        return id;
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public String getCorreo() {
+        return correo;
+    }
+
+    public String getDni() {
+        return dni;
+    }
+
+    public String getContrasenia() {
+        return contrasenia;
+    }
+
+    public String getRol() {
+        return rol;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+    public void setCorreo(String correo) {
+        this.correo = correo;
+    }
+
+    public void setDni(String dni) {
+        this.dni = dni;
+    }
+
+    public void setContrasenia(String contrasenia) {
+        this.contrasenia = contrasenia;
+    }
+
+    public void setRol(String rol) {
+        this.rol = rol;
+    }
+}
